Stop player kicked out of vehicle by char. change
This is the only real solution, because getting kicked out of one's
vehicle when changing player character is intended behaviour (otherwise
the character entity would be left in an invalid state).

It occurs to me that the vehicle would also be impossible to enter for
no apparent reason (no feedback that a character entity is inside), so
this is almost certainly a good fix for us.

Closes #66.