define new on_updated event
on_updated is raised after update_forces completes and sends the amount, ore per minute, and permille remaining for each site. the data is keyed by force and site name.

get_on_updated_event_id is added to the remote interface to allow other mods to discover the event id and subscribe to it