Use a separate `ore_tracker` module
Instead of having ores be referenced by the sites themselves, and instead of
requiring a `surface.find_entity()` call every time we need to update an ore's
remaining quantity, we leave the tracking of ore to a dedicated library, and
simply pull data from a table when we need updates.

Among other things, this allows us a central place that reads a fixed number of
entities each tick (closes #47), doesn't need to perform a (costly)
`find_entity` periodically, and avoids double-updating entities that are in two
different sites.

Closes #98.