Optionally group by ore type
If per-player setting YARM-group-by-ore is checked, sort resource list by ore, THEN percentage remaining.
I was unable to add a third parameter to the comparison function used in the table.sort call, and therefore unable to pass YARM-group-by-ore to within site_comparator (either directly, or via player object reference).  Instead, I check YARM-group-by-ore in ascending_by_ratio and choose which table.sort call to use depending on the setting.  If there's a better way to do this, I'd love to learn about it.